refactor: [DefinitionsConfigurator] Move the definitions configurator to a separate class.
