+++ tests/Attributes/MultiAutowireTest.php
@@ -6,6 +6,7 @@ namespace Tests\Attributes;
 
 use Kaspi\DiContainer\AttributeReader;
 use Kaspi\DiContainer\Attributes\Autowire;
+use Kaspi\DiContainer\DefinitionsConfigurator;
 use Kaspi\DiContainer\DefinitionsLoader;
 use Kaspi\DiContainer\DiContainer;
 use Kaspi\DiContainer\DiContainerBuilder;
@@ -53,6 +54,7 @@ use Tests\Attributes\Fixtures\MultiAutowire;
 #[CoversClass(SourceDefinitionItem::class)]
 #[CoversClass(FreezeTrait::class)]
 #[CoversClass(EventListener::class)]
+#[CoversClass(DefinitionsConfigurator::class)]
 class MultiAutowireTest extends TestCase
 {
     public function testMultiAutowireContainer(): void
