+++ tests/Compiler/CompilableDefinition/ObjectEntry/CompiledContainerWIthObjectResettersTest.php
@@ -17,6 +17,7 @@ use Kaspi\DiContainer\Compiler\DiContainerDefinitions;
 use Kaspi\DiContainer\Compiler\DiDefinitionTransformer;
 use Kaspi\DiContainer\Compiler\Helper;
 use Kaspi\DiContainer\Compiler\IdsIterator;
+use Kaspi\DiContainer\DefinitionsConfigurator;
 use Kaspi\DiContainer\DefinitionsLoader;
 use Kaspi\DiContainer\DiContainer;
 use Kaspi\DiContainer\DiContainerBuilder;
@@ -92,6 +93,7 @@ use function random_bytes;
 #[CoversClass(ContainerParametersEntry::class)]
 #[CoversClass(AbstractSourceParameters::class)]
 #[CoversClass(EventListener::class)]
+#[CoversClass(DefinitionsConfigurator::class)]
 class CompiledContainerWIthObjectResettersTest extends TestCase
 {
     public function testObjectResettersCompilation(): void
