+++ tests/Compiler/CompileLazyObject/CompileLazyObjectTest.php
@@ -18,6 +18,7 @@ use Kaspi\DiContainer\Compiler\DiContainerDefinitions;
 use Kaspi\DiContainer\Compiler\DiDefinitionTransformer;
 use Kaspi\DiContainer\Compiler\Helper;
 use Kaspi\DiContainer\Compiler\IdsIterator;
+use Kaspi\DiContainer\DefinitionsConfigurator;
 use Kaspi\DiContainer\DefinitionsLoader;
 use Kaspi\DiContainer\DiContainer;
 use Kaspi\DiContainer\DiContainerBuilder;
@@ -95,6 +96,7 @@ use function random_bytes;
 #[CoversClass(SetupAttributeTrait::class)]
 #[CoversClass(ValueEntry::class)]
 #[CoversClass(EventListener::class)]
+#[CoversClass(DefinitionsConfigurator::class)]
 class CompileLazyObjectTest extends TestCase
 {
     public function testCompileLazyObject(): void
