+++ tests/ContainerBuilder/DiContainerBuilderConfiguratorContextTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Tests\ContainerBuilder;
 
+use Kaspi\DiContainer\DefinitionsConfigurator;
 use Kaspi\DiContainer\DefinitionsLoader;
 use Kaspi\DiContainer\DiContainer;
 use Kaspi\DiContainer\DiContainerBuilder;
@@ -41,6 +42,7 @@ use PHPUnit\Framework\TestCase;
 #[CoversFunction('Kaspi\DiContainer\diAutowire')]
 #[CoversClass(FreezeTrait::class)]
 #[CoversClass(EventListener::class)]
+#[CoversClass(DefinitionsConfigurator::class)]
 class DiContainerBuilderConfiguratorContextTest extends TestCase
 {
     public function testSetContexts(): void
