+++ tests/DefinitionsLoader/DefinitionsLoaderWithConfiguratorTest.php
@@ -7,6 +7,7 @@ namespace Tests\DefinitionsLoader;
 use Kaspi\DiContainer\AttributeReader;
 use Kaspi\DiContainer\Attributes\Autowire;
 use Kaspi\DiContainer\Attributes\Tag;
+use Kaspi\DiContainer\DefinitionsConfigurator;
 use Kaspi\DiContainer\DefinitionsLoader;
 use Kaspi\DiContainer\DiDefinition\DiDefinitionAutowire;
 use Kaspi\DiContainer\DiDefinition\DiDefinitionTaggedAs;
@@ -59,6 +60,7 @@ use function Kaspi\DiContainer\diValue;
 #[CoversClass(ResetterTrait::class)]
 #[CoversFunction('\Kaspi\DiContainer\diValue')]
 #[CoversClass(EventListener::class)]
+#[CoversClass(DefinitionsConfigurator::class)]
 class DefinitionsLoaderWithConfiguratorTest extends TestCase
 {
     public function testCircularLoadFromFile(): void
