+++ tests/DiContainer/RemovedDefinitionIds/RemovedDefinitionIdsTest.php
@@ -16,6 +16,7 @@ use Kaspi\DiContainer\Compiler\ContainerCompilerToFile;
 use Kaspi\DiContainer\Compiler\DiContainerDefinitions;
 use Kaspi\DiContainer\Compiler\DiDefinitionTransformer;
 use Kaspi\DiContainer\Compiler\IdsIterator;
+use Kaspi\DiContainer\DefinitionsConfigurator;
 use Kaspi\DiContainer\DefinitionsLoader;
 use Kaspi\DiContainer\DiContainer;
 use Kaspi\DiContainer\DiContainerBuilder;
@@ -86,6 +87,7 @@ use function random_bytes;
 #[CoversClass(ContainerParametersEntry::class)]
 #[CoversClass(AbstractSourceParameters::class)]
 #[CoversClass(EventListener::class)]
+#[CoversClass(DefinitionsConfigurator::class)]
 class RemovedDefinitionIdsTest extends TestCase
 {
     public function testRemovedDefinitionIds(): void
